building/trading_core/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
python-dotenv>=0.19.0

# Optional dependencies
shioaji>=0.3.0  # Sinopac broker API
matplotlib>=3.4.0  # For plotting
scikit-learn>=0.24.0  # For linear regression and ML tools

# Development
pytest>=6.0.0
black>=21.0
mypy>=0.910
flake8>=3.9.0
